refactor: getJsonPathForPosition and getLocationForJsonPath
