+++ src/pathToPointer.ts
@@ -1,11 +1,11 @@
-import { JSONPath } from '@stoplight/types';
+import { JsonPath } from '@stoplight/types';
 import { encodePointerFragment } from './encodePointerFragment';
 
-export const pathToPointer = (path: JSONPath): string => {
+export const pathToPointer = (path: JsonPath): string => {
   return encodeUriFragmentIdentifier(path);
 };
 
-const encodeUriFragmentIdentifier = (path: JSONPath): string => {
+const encodeUriFragmentIdentifier = (path: JsonPath): string => {
   if (path && typeof path !== 'object') {
     throw new TypeError('Invalid type: path must be an array of segments.');
   }
